Ensuring update sidenotes render
Prior to this commit, with Hugo v0.58.0 and greater, I was
encountering situations in which the sidenote content (eg. the INS)
was not rendering on the side. That was due to a double class
declaration.

The second problem was that the content after an update sidenote was
rendered outside of a P-tag. This addresses that by compressing the
shortcode rendering into single lines.