Relocate relationships.html to help webmentions.io
I think webmentions.io looks at the first occurence of an outgoing link for any
microformats, and that /was/ happening in a footer that appeared after any use
of the link in the body of the post. Which is not good. So, I've moved the
relationships partial into summary-section.html, which appears before the content.